Update valid_string.java
given a string consisting of characters 'a' and 'b'. Your task is to determine whether the string is valid based on the following rules:
The string can start with 'a'.
Each 'a' can be followed by either another 'a' or two consecutive 'b's.
If the string starts with 'b' or has any other invalid pattern, it should be considered invalid.